Revert "Community-contributed localization for Afrikaans (af)"
This reverts commit b354bca00007496272edf844129e144ea571658d.

Unfortunately, Afrikaaans is the default selection of the language, and
some people apparently contributed translations without selecting the
proper locale. I see some Hebrew text in here, which was easier to
distinguish because it uses an entirely different character set.

But I have to assume other locales are mixed in here as well.

I'll remove the default language from the translation assistance plugin.
But in the mean time, we don't have anyone who's capable of reviewing
this, so I'm forced to roll back.