Fixed test failures on core versions >= 1.597
This change also fixes hpi:run not having the plugin's extensions registered
due to circular dependencies.

Content of the descriptions field of AbstractTestResultAction caused old builds
not to be fully populated during the conversion so for example the status was kept as FAILURE
instead of SUCCESS when read from disk. As the tests don't want to test that part
it was a simple operation of removing that field from the build.xml so that the tests
had correct data to use.