Testsuite duration may be parsed from time attr
Testsuite duration can be retrieved from its time attribute if
it exists. Otherwise it will continue to be calculated from the sum
of it's testcases. ClassResult duration is now obtained from it's
cases' SuiteResult since it was just re-summing case durations. This
follows the JUNIT XML Standard which includes the time attributed
in testsuites as well as test cases.