Allow to print Failure and Error messages together
It happens when running tests with pytest that we can have a
test failing with both failure and error.

ex:
import pytest
import logging

@pytest.fixture()
def setup_class(request):
    def finalize():
        raise Exception("test teadown Exception")
    request.addfinalizer(finalize)

def test_1(setup_class):
    assert 1==2

In this case the XML report will contain two nodes, a failure
and an error and both messages are usefull in the jenkins TestReport.
Currently, the plugin chooses to display the error report over
the failure. This patch allows failure message to appear as well.