JENKINS-42438 Correctly calculate test class duration.
When running Android tests the <testsuite> in the generated xml file
can contain multiple classes.
Thus setting the test class duration to the value of the test-suite is wrong.

Instead the class duration is now calculated as sum its test cases.
This is the same way it was done before #35 was merged.
Reverting parts of both #35 and #54 did not break any existing tests.