[FIXED JENKINS-48196] Remove redundant branch names from case displays
If we've got a suite with a stage inside a parallel branch where the
stage and parallel branch have the same name, we should only include
that name once in the suite/case/etc display name - it's redundant to
display it twice.

There is a potential gotcha here if you've got multiple stages inside
a branch and one of those stages has the same name as the branch, but
I think that's an edge case we can live with for now.