[JENKINS-53365] do not exclude transitive plugin deps
excluding a transative plugin dependency is a recepie for disaster for
consumers of the plugin.

It will cause anyone who is depending on this plugin to have to declare
a dependency on the excluded plugin in order for the InjectedTest or
hpi:run to work.

Given that the optionality was removed earlier I am just removing the
exclusion, however I do not fully understand why the optionality on the
workflow-api could not have been retained.