Disable automatic test file validation and add validate button.
Automatic test file validation causes issues when workspaces have
large O(millions) amount of files. In those cases, if the particular
test file doesn't exist at the time the user looks at the
configuration page, the Agent will start crawling the whole workspace
looking for other subdirectories where that file might be.

By doing so, it's tracking all the files it's seen which eventually
leads to filling up the heap and crashing.

This disables the automatic file validation and adds a button that
makes it possible to validate manually.