potential race condition in result aggregation
In attempting to diagnose an ATH failure I noticed what looks like a
potential race condition that could lead to results not being correctly
updated.

For example in rare cases (but maybe not so rare in our ATH) if 2 builds finish at
T1, a web request comes in a T2 and then another build fishes at T3
followed by a web request at T4.

When T1 arrives the timestamp is set to T1, and then in T2 the current
value is checked against the timestamp and set to T1+1

When T3 arrives the timestamp is set to T3.

THe issue occurs if you have multiple CPUs and updating the code is fast
and T1 = T2 = T3.

When the request at T4 arrives the code thinks everything is up-to-date when it is not.