If the jobs are not specified we need to calculate them on demand
The dependencies may not be calculated when we create the action.

The dependency graph is updated asynchronously by Jenkins, so we may not
have the full graph when the TestResultAction is created

This caused the job calculation to be fixed in stone, and potentially
incorrect if the test results where obtained before the graph calculation had completed.

By deferring obtaining the downstream jobs we reduce the risk of partial
results being shown, and they will be eventually complete (when some run
completes).
We are not computing the graph so this should have little overhead on
the action.